journalEntryのバックエンド完了

+++ src/main/kotlin/com/example/accounting/domain/journalEntry/JournalEntryAmount.kt
@@ -1,17 +1,17 @@
 package com.example.accounting.domain.journalEntry
 
 class JournalEntryAmount(
-    val value: Long,
+    val value: Int,
 ) {
     companion object {
 
 
-        fun of(value: Long): JournalEntryAmount {
+        fun of(value: Int): JournalEntryAmount {
             validate(value)
             return JournalEntryAmount(value)
         }
 
-        private fun validate(value: Long) {
+        private fun validate(value: Int) {
             if (value <= 0) {
                 throw RuntimeException("金額は0以上の数字で入力してください")
             }
